code/requirements.txt
numpy==1.16.4
numba==0.48
pytorch-lightning==0.7.6
Cython==0.29.20
asteroid==0.3.0
sagemaker-inference==1.3.2.post0
PyYAML==5.3.1